requirements.txt
fastapi
uvicorn
pydantic
websockets
paho-mqtt>=2.0.0
